+++ UniPlaySongSettingsView.xaml.cs
@@ -92,6 +92,8 @@ namespace UniPlaySong
             s.GamePropFilterGenreIds = new List<Guid>();
             s.GamePropFilterSourceIds = new List<Guid>();
             s.FilterModeEnabled = false;
+            s.RadioModeEnabled = false;
+            s.RadioMusicSource = RadioMusicSource.FullLibrary;
             s.EnableCompletionCelebration = true;
             s.CelebrationSoundType = CelebrationSoundType.BundledJingle;
             s.SelectedCelebrationJingle = "Streets of Rage 1 - Sega Genesis - Level Clear.mp3";